MQTTsend modifications
MQTT is now enabled by default and the MQTT send interval is set to be 5 minutes. The device sends an MQTT message as soon as it has connected to MQTT.
Diagnostic messages have been improved.
Note that the power output control is under test and presently turns the power output on and off at 1Hz.